refactor(sf): Extract common Salesforce connector logic to base class
- Created BaseSalesforceConnector class to reduce code duplication
- Extracted parse_salesforce_timestamp function to base class
- Moved OAuth and legacy authentication logic to base class
- Moved query retry logic to base class
- Updated all Salesforce connectors to inherit from BaseSalesforceConnector
- Removed manual pointer updates (now handled by BaseConnector.save())
- Fixed data ordering for setup_audit_trail (reverses DESC query results)
- Updated test files to patch base class instead of individual connectors
- All tests passing, linting and type checking pass
